- Removed levelTEMP.png as it is now unused + Added level1.png as a test-level creation (This will be re-done eventually) + Added dirt.png to represent a dirt texture (This is a temporary texture, Jenna Young will be assisting me with texturing) + Added new (unused) Gravity interface to handle gravity integers * Renamed "ImageLoader" to "ImageUtils" - Removed (unused) TileGenerator
TileType:
- Removed previous tile types
+ Added DIRT, GRASS, and STONE tile types
+ Added static getById(int id) and static getByName(String name) methods to get a tile by respective method parameters

Location:
+ Added a new public static final PIXEL_RATIO integer to determine how large the pixel:coordinate ratio is
* x and y are now float based rather than integer based
* Modified setX(), getX(), setY(), and getY() to work with a new coordinate system (32 pixels = 1 coordinate)
+ Added setRawX(), getRawX(), setRawY(), and getRawY() methods to set the coordinates by pixels (Used in entity classes)
+ Added clone() method to create a new Location object based on the current coordinates
+ Added distance() method to check the distance between two points
+ Overrided the equals() method to properly check if two locations are similar

Level:
+ Added private "tileSize" integer field to determine how large tile sizes should be
* Changed from 2dimensional array to Map<Location, Tile> to keep track of tiles
* Modified getTileAt(), setTile(), and other methods to work with the new tile Map
* Fixed the tick() method not having a proper method body (Oops)

Tile:
* Changed "image" to "texture" respectively
* Made changes to reflect the Location changes for the new coordinate system
- Removed final modifier from "type" field
- Removed "loaded" field, and its associate methods
+ Added private static final Camera object to properly render tiles relative to the camera offset
+ Added setTileType(TileType) method to change a tiles type mid-level
+ Added a toString method for debug purposes

Texture:
* Spaced out the enums to be more organized
- Removed GAME_LEVELTEMP enum due to no use
+ Added GAME_TILE_DIRT enum with its respective texture

ImageUtils:
+ Added getPixelColor(int) method to get a Color value from an RGB pixel
+ Added isColor(int pixel, int red, int green, int blue) and isColor(int pixel, Color color) methods to easily check if pixel colors are similar

LevelManager:
+ Added a method body for loadLevel() which will read each pixel in a file and convert it into a level
- Removed saveLevel() method
+ Added setCurrentLevel(int index) and setCurrentLevel(Level) methods

EntityHandler:
- Removed "tile collision" which was honestly justa wreck. What was I thinking when I wrote this. TODO: Redo it
+ Added new private DecimalFormat field to set new float values as
properly formatted decimals
* Modified the debug x and y coordinate displays respectively

GameObject:
+ Added a protected static final Random field to allow for randomization
throughout game objects

Player:
+ Added a speed integer to allow for an easy change of speed for most
features such as camera movement, and player velocity changes
* Changed the width/height of player from 30 to 32
* Made modifications to reflect changes in the Location object and new coordinate system
+ Let Player implement the new Gravity class and returned a temporary
value of 1

Entity:
* Velocity values are now floats rather than integers to be more precise with the new coordinate system

Game:
* Changed "handler" field to "entityHandler" for clarification
* Changed currentFPS to be public and static
+ Loaded the first level (and set it as the current level) in the init
method
* Temporarily removed the Enemy object while working with levels

+++ src/me/choco/game/entity/Enemy.java
@@ -25,6 +25,6 @@ public class Enemy extends Entity{
 	@Override
 	public void render(Graphics g) {
 		g.setColor(Color.RED);
-		g.fillRect(location.getX() - camera.getXOffset(), location.getY() - camera.getYOffset(), width, height);
+		g.fillRect((int) location.getX() - camera.getXOffset(), (int) location.getY() - camera.getYOffset(), width, height);
 	}
 }
